Several smaller fixes and changes
#1230
#1229
#1228
#1226

+++ ingamescriptsupport/events/scripteventchangecobar.cpp
@@ -128,7 +128,7 @@ void ScriptEventChangeCOBar::showEditEvent(spScriptEditor pScriptEditor)
     connect(spinBox.get(), &SpinBox::sigValueChanged, this,
             [this](qreal value)
     {
-        setStars(static_cast<qint32>(value));
+        setStars(value);
     });
     pBox->addItem(spinBox);
